chore: modules proposal
