+++ apps/web/pages/bookings/[status].tsx
@@ -1,0 +1,16 @@
+import type { GetStaticPaths } from "next";
+
+import { validStatuses } from "~/bookings/lib/validStatuses";
+
+export { getStaticProps } from "~/bookings/views/bookings-listing-view.getStaticProps";
+export { default } from "~/bookings/views/bookings-listing-view";
+
+export const getStaticPaths: GetStaticPaths = () => {
+  return {
+    paths: validStatuses.map((status) => ({
+      params: { status },
+      locale: "en",
+    })),
+    fallback: "blocking",
+  };
+};
